refactor: 상수, 도메인 패키지 분리
