+++ src/main/java/lotto/constants/LottoConstants.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.constants;
 
 public class LottoConstants {
     public static final int LOTTO_NUMBER_COUNT = 6;
