+++ src/main/java/lotto/constants/LottoMachineConstants.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.constants;
 
 public class LottoMachineConstants {
     public static final int LOTTO_PRICE = 1000;
